% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyformula.R
\name{tidyformula}
\alias{tidyformula}
\title{Build a formula using \code{tidyselect}-style selection helperts}
\usage{
tidyformula(
  formula,
  data,
  select_helpers = .select_helpers,
  env = rlang::caller_env()
)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}. Can contain selection helpers
to be expanded}

\item{select_helpers}{A character vector. The names of selection helpers to
be matched and substituted}

\item{env}{The environment to associate with the result}

\item{df}{A data frame whose column names should be used for selection}
}
\description{
\code{tidyformula} translate formulas containing \code{tidyselect}-style
\link[tidyselect:language]{selection helpers}, expanding these helpers by evaluating
\code{\link[dplyr:select]{dplyr::select}} with the relevant selection helper and a supplied data frame.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{df1 <- data.frame(
  x1 = rnorm(5),
  x2 = rnorm(5),
  y  = rnorm(5)
)

tidyformula(y ~ starts_with("x") + z, data = df1)
#> y ~ x1 + x2 + z
#> <environment: 0x000002b40112c830>
}\if{html}{\out{</div>}}
}

